[#182] Refactor: useProfileBox 커스텀 훅
