+++ components/MyPage/ProfileBox.tsx
@@ -1,12 +1,10 @@
-import API from '@/apis/api';
 import AddImageButton from '@/components/AddImageButton/AddImageButton';
 import BasicInput from '@/components/Input/ModalInputContainer/BasicInput';
 import Button from '@/components/common/Button/Button';
-import { IMG_URL_ERROR, NICKNAME_ERROR, NICKNAME_IMG_ERROR } from '@/constants/ErrorMsg';
-import useUserData from '@/hooks/global/useUserData';
+import useProfileBox from '@/hooks/MyPage/useProfileBox';
 import { onMobile, onTablet } from '@/styles/mediaQuery';
 
-import { Dispatch, SetStateAction, useState } from 'react';
+import { Dispatch, SetStateAction } from 'react';
 import { styled } from 'styled-components';
 
 interface Props {
@@ -18,69 +16,7 @@ interface Props {
 }
 
 function ProfileBox({ email, nickname, profileImg, setNickName, setPreviewImage }: Props) {
-  const { setUser, user } = useUserData();
-  const [image, setImage] = useState<File>();
-  const [errorMsg, setErrorMsg] = useState('');
-
-  const changeProfile = async () => {
-    if (!image) {
-      try {
-        await API.users.correctMyInfo({ nickname });
-        alert('프로필 업데이트 완료 🍀');
-        user.nickname = nickname;
-        setUser(user);
-      } catch (e: any) {
-        switch (e.data.message) {
-          case NICKNAME_ERROR:
-            setErrorMsg(NICKNAME_ERROR);
-            break;
-          case NICKNAME_IMG_ERROR:
-            setErrorMsg(NICKNAME_IMG_ERROR);
-            break;
-          case IMG_URL_ERROR:
-            setErrorMsg(IMG_URL_ERROR);
-            break;
-          default:
-            setErrorMsg(e.data.message);
-            break;
-        }
-      }
-
-      return;
-    } else {
-      const formData = new FormData();
-      formData.append('image', image);
-
-      if (image.type !== 'image/svg+xml') {
-        const response = await API.users.profileImgUpload(formData);
-        const { profileImageUrl } = response;
-        try {
-          await API.users.correctMyInfo({ profileImageUrl, nickname });
-          alert('프로필 업데이트 완료 🍀');
-          user.profileImageUrl = profileImageUrl;
-          user.nickname = nickname;
-          setUser(user);
-        } catch (e: any) {
-          switch (e.data.message) {
-            case NICKNAME_ERROR:
-              setErrorMsg(NICKNAME_ERROR);
-              break;
-            case NICKNAME_IMG_ERROR:
-              setErrorMsg(NICKNAME_IMG_ERROR);
-              break;
-            case IMG_URL_ERROR:
-              setErrorMsg(IMG_URL_ERROR);
-              break;
-            default:
-              setErrorMsg(e.data.message);
-              break;
-          }
-        }
-      } else {
-        alert('형식에 맞는 이미지만 사용가능합니다.');
-      }
-    }
-  };
+  const { changeProfile, errorMsg, setImage } = useProfileBox({ nickname });
 
   return (
     <StyledContainer>
